refactor: split extract into a directory, fix empty package tarball
extract.ts split along the pipeline: runs (types), walk (DOM to raw runs),
collapse (whitespace collapsing), wrappers (inline box geometry), index
(entry point). Import paths are unchanged.

npm pack shipped src/ and no dist/, since dist is gitignored and the
manifest declared no files field. Add it, and assert the packed shape
rather than the publishable shape the package no longer has.

Add SUMMARY and SEARCH to the clipboard block-tag set; DETAILS was
already there but its own SUMMARY was not.
