+++ packages/linebreak/package.json
@@ -7,6 +7,9 @@
   "packageManager": "bun@1.3.14",
   "license": "MIT",
   "author": "jktrn",
+  "files": [
+    "dist"
+  ],
   "sideEffects": [
     "./dist/styles.css"
   ],
