+++ packages/linebreak/src/dom/clipboard.ts
@@ -31,7 +31,9 @@ const BLOCK_TAGS = new Set([
   "OL",
   "P",
   "PRE",
+  "SEARCH",
   "SECTION",
+  "SUMMARY",
   "TABLE",
   "TD",
   "TH",
